Move all filterlist components to the same file
