+++ src/panels/SettingsPanel/SettingsPanel.tsx
@@ -1,16 +1,25 @@
-import { useMemo } from 'react';
+import { useCallback, useMemo } from 'react';
 import { Container, ScrollArea } from '@mantine/core';
 
+import { useGetOptionPropertyValue } from '@/api/hooks';
 import { FilterList } from '@/components/FilterList/FilterList';
 import { PropertyOwner } from '@/components/PropertyOwner/PropertyOwner';
 import { useAppSelector } from '@/redux/hooks';
-import { ScenePrefixKey } from '@/util/keys';
-import { isTopLevelPropertyOwner } from '@/util/propertyTreeHelpers';
+import { EnginePropertyVisibilityKey, ScenePrefixKey } from '@/util/keys';
+import {
+  hasVisibleChildren,
+  identifierFromUri,
+  isTopLevelPropertyOwner
+} from '@/util/propertyTreeHelpers';
+import { checkCaseInsensitiveSubstringList } from '@/util/stringmatcher';
 
-import { SettingsSearchList } from './SettingsSearchList';
+import { SettingsSearchListItem } from './SettingsSearchListItem';
+import { collectSearchableItems, SearchItem, SearchItemType } from './util';
 
 export function SettingsPanel() {
   const propertyOwners = useAppSelector((state) => state.propertyOwners.propertyOwners);
+  const propertyOverview = useAppSelector((state) => state.properties.propertyOverview);
+  const [visiblityLevelSetting] = useGetOptionPropertyValue(EnginePropertyVisibilityKey);
 
   // Get all the top property owners, that are not part of the scene
   const topLevelPropertyOwners = useMemo(
@@ -20,6 +29,45 @@ export function SettingsPanel() {
       ),
     [propertyOwners]
   );
+
+  const renderfunc = useCallback(
+    (item: SearchItem) => (
+      <SettingsSearchListItem key={item.uri} type={item.type} uri={item.uri} />
+    ),
+    []
+  );
+
+  const matcher = useCallback((testItem: SearchItem, search: string): boolean => {
+    return checkCaseInsensitiveSubstringList(testItem.searchKeys, search);
+  }, []);
+
+  const searchData = useMemo(() => {
+    let [searchableSubOwners, searchableProperties] = collectSearchableItems(
+      topLevelPropertyOwners,
+      propertyOwners
+    );
+    searchableSubOwners = searchableSubOwners.filter((uri) =>
+      hasVisibleChildren(uri, visiblityLevelSetting, propertyOwners, propertyOverview)
+    );
+    searchableProperties = searchableProperties.filter(
+      (uri) =>
+        visiblityLevelSetting && visiblityLevelSetting >= propertyOverview[uri].visibility
+    );
+
+    return searchableSubOwners
+      .map((uri) => ({
+        type: SearchItemType.SubPropertyOwner,
+        uri,
+        searchKeys: [propertyOwners[uri]!.name, identifierFromUri(uri)]
+      }))
+      .concat(
+        searchableProperties.map((uri) => ({
+          type: SearchItemType.Property,
+          uri,
+          searchKeys: [propertyOverview[uri].name, identifierFromUri(uri)]
+        }))
+      );
+  }, [propertyOverview, propertyOwners, topLevelPropertyOwners, visiblityLevelSetting]);
 
   return (
     <ScrollArea h={'100%'}>
@@ -31,8 +79,15 @@ export function SettingsPanel() {
               <PropertyOwner uri={uri} key={uri} />
             ))}
           </FilterList.Favorites>
-          {/* This component creates the FilterList.Data component */}
-          <SettingsSearchList searchableTopOwners={topLevelPropertyOwners} />
+          <FilterList.Data<SearchItem>
+            data={searchData}
+            // Do not virtualize here, since this requires the rendered objects to have a
+            // fixed size, which PropertyOwners won't. Also, the list is not expected to
+            // be very long
+            virtualize={false}
+            renderElement={renderfunc}
+            matcherFunc={matcher}
+          />
         </FilterList>
       </Container>
     </ScrollArea>
